backend/requirements.txt
Django>=4.2.0

djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-redis>=5.2.0
redis>=4.5.0
celery>=5.3.0
channels>=4.0.0
channels-redis>=4.1.0
Pillow>=10.0.0